+++ Runtime/Singleton/MonoSingleton.cs
@@ -22,7 +22,7 @@ namespace AchEngine
             {
                 if (instance == null)
                 {
-                    instance = FindObjectOfType<T>();
+                    instance = FindFirstObjectByType<T>();
                     if (instance == null)
                     {
                         var obj = new GameObject(typeof(T).Name);